feat. anpassung moved test classes to one file, and cleared some of louis' comments
